split fileio into utils, mypy checkpoint
