+++ src/zenml/core/base_component.py
@@ -18,6 +18,7 @@ from uuid import UUID, uuid4
 
 from pydantic import BaseSettings, Field, root_validator
 
+import zenml.io.utils
 from zenml.core.utils import generate_customise_sources
 from zenml.io import fileio
 from zenml.logger import get_logger
@@ -98,7 +99,7 @@ class BaseComponent(BaseSettings):
             sort_keys=True,
             exclude={SUPERFLUOUS_OPTIONS_ATTRIBUTE_NAME},
         )
-        fileio.write_file_contents_as_string(file_path, file_content)
+        zenml.io.utils.write_file_contents_as_string(file_path, file_content)
 
     def _create_serialization_file_if_not_exists(self) -> None:
         """Creates the serialization file if it does not exist."""
