+++ src/zenml/core/local_service.py
@@ -1,10 +1,10 @@
 from typing import TYPE_CHECKING, Dict
 
+import zenml.io.utils
 from zenml.core import mapping_utils
 from zenml.core.base_component import BaseComponent
 from zenml.core.mapping_utils import UUIDSourceTuple
 from zenml.exceptions import AlreadyExistsException, DoesNotExistException
-from zenml.io import fileio
 from zenml.logger import get_logger
 from zenml.stacks.base_stack import BaseStack
 from zenml.utils import source_utils
@@ -44,7 +44,7 @@ class LocalService(BaseComponent):
 
     def get_serialization_dir(self) -> str:
         """The local service stores everything in the zenml config dir."""
-        return fileio.get_zenml_config_dir()
+        return zenml.io.utils.get_zenml_config_dir()
 
     def get_serialization_file_name(self) -> str:
         """Return the name of the file where object is serialized."""
